backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.49
alembic==1.13.1
pydantic
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
passlib==1.7.4
python-multipart==0.0.6
email-validator==2.3.0
httpx==0.26.0
openpyxl==3.1.2
pandas=3.0.2
reportlab==4.0.8
celery==5.3.6
redis==5.0.1
python-dotenv==1.0.0
aiosmtplib==3.0.1
